requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
langchain-text-splitters==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
unstructured==0.15.12
unstructured-client==0.25.8
unstructured-inference==0.7.36
unstructured[pdf]
faiss-gpu==1.7.2
sentence-transformers==3.1.0
nltk==3.9.1
transformers==4.44.2
docx2txt==0.8
python-docx==1.1.2
openpyxl==3.1.5
ocrmypdf==16.5.0
poppler-utils==0.1.0